AUTO UPDATE mode for enc_decode_sav
When in AUTO UPDATE mode the enc_decode_sav utility waits for changes in the selected file and immediately converts it for as long as you want.

Useful when experimenting with a SAV.json file: you edit and save it and then immediately load the corresponding SAV to the game (the enc_decode_sav in AUTO UPDATE mode automatically converts SAV.json to SAV when you save)

And no more sav.enc files: backup files (.bak) are created instead.